refactor: Crear módulo de beneficios. #230
