+++ src/benefit/components/benefitCardContent/BenefitCardContent.module.css
@@ -1,4 +1,4 @@
-@import "../../../../../shared/styles/colors.css";
+@import "../../../shared/styles/colors.css";
 
 .contentContainer {
   display: flex;
