+++ src/benefit/types/benefit.types.ts
@@ -1,90 +1,7 @@
 import type { SubjectSimplifiedResponseDto } from "../../admin/services/subject/SubjectService";
 import type { PaginatedData } from "../../shared/types/PaginacionType";
 
-type BenefitStudentState =
-  | "AVAILABLE"
-  | "PURCHASED"
-  | "USE_REQUESTED"
-  | "EXPIRED";
-
-export interface BenefitResponseInterface {
-  id: number;
-  name: string;
-  description: string;
-  cost: number;
-  purchaseLimit: number | null;
-  purchaseLimitPerStudent: number | null;
-  subjectDto: SubjectSimplifiedResponseDto;
-  color: Color;
-  endAt: string;
-  icon: Icon;
-  category: Category;
-}
-
-export interface BenefitStudentResponseInterface {
-  id: number;
-  name: string;
-  description: string;
-  cost: number;
-  state: BenefitStudentState;
-  purchasesLeft: number | null;
-  purchasesLeftByStudent: number | null;
-  endAt: string;
-  subjectId: number;
-  subjectName: string;
-  icon: Icon;
-  category: Category;
-  color: Color;
-}
-
-export interface PaginatedBenefitStudentResponseInterface {
-  data: PaginatedData<BenefitStudentResponseInterface>;
-  message: string;
-  errors: any;
-  timestamp: string;
-}
-
-export interface BenefitStatsResponse {
-  available: number;
-  purchased: number;
-  use_requested: number;
-  used: number;
-  expired: number;
-}
-
-export interface BenefitStatsApiResponse {
-  data: BenefitStatsResponse;
-  message: string;
-  errors: any;
-  timestamp: string;
-}
-
-export interface CreateBenefitInterface {
-  name: string;
-  description: string;
-  cost: number | string;
-  purchaseLimit: number | null;
-  purchaseLimitPerStudent: number | null;
-  purchaseLeft?: number;
-  subjectId: number;
-  endAt: string;
-  color: Color;
-  icon: Icon;
-  category: Category;
-}
-
-export interface BenefitValidationErrors {
-  name?: string;
-  description?: string;
-  cost?: string;
-  endAt?: string;
-  subjectId?: string;
-  icon?: string;
-  category?: string;
-  color?: string;
-  purchaseLimit?: string;
-  purchaseLimitPerStudent?: string;
-}
+// ==================== ENUMS Y CONSTANTES ====================
 
 export const colors = [
   "BLUE",
@@ -118,3 +35,108 @@ export const categories = [
 export type Color = (typeof colors)[number];
 export type Icon = (typeof icons)[number];
 export type Category = (typeof categories)[number];
+
+export type BenefitStudentState =
+  | "AVAILABLE"
+  | "PURCHASED"
+  | "USE_REQUESTED"
+  | "EXPIRED";
+
+export type BenefitVariant = "student" | "teacher";
+
+// ==================== INTERFACES DE RESPUESTA ====================
+
+export interface BenefitResponseInterface {
+  id: number;
+  name: string;
+  description: string;
+  cost: number;
+  purchaseLimit: number | null;
+  purchaseLimitPerStudent: number | null;
+  subjectDto: SubjectSimplifiedResponseDto;
+  color: Color;
+  endAt: string;
+  icon: Icon;
+  category: Category;
+}
+
+export interface BenefitStudentResponseInterface {
+  id: number;
+  name: string;
+  description: string;
+  cost: number;
+  state: BenefitStudentState;
+  purchasesLeft: number | null;
+  purchasesLeftByStudent: number | null;
+  endAt: string;
+  subjectId: number;
+  subjectName: string;
+  icon: Icon;
+  category: Category;
+  color: Color;
+}
+
+export interface CreateBenefitInterface {
+  name: string;
+  description: string;
+  cost: number | string;
+  purchaseLimit: number | null;
+  purchaseLimitPerStudent: number | null;
+  purchaseLeft?: number;
+  subjectId: number;
+  endAt: string;
+  color: Color;
+  icon: Icon;
+  category: Category;
+}
+
+// ==================== TIPOS UTILITARIOS ====================
+
+export type AnyBenefit =
+  | BenefitResponseInterface
+  | BenefitStudentResponseInterface
+  | CreateBenefitInterface;
+
+// ==================== PAGINACIÓN Y STATS ====================
+
+export interface PaginatedBenefitStudentResponseInterface {
+  data: PaginatedData<BenefitStudentResponseInterface>;
+  message: string;
+  errors: any;
+  timestamp: string;
+}
+
+export interface BenefitStatsResponse {
+  available: number;
+  purchased: number;
+  use_requested: number;
+  used: number;
+  expired: number;
+}
+
+export interface BenefitStatsApiResponse {
+  data: BenefitStatsResponse;
+  message: string;
+  errors: any;
+  timestamp: string;
+}
+
+// ==================== VALIDACIÓN ====================
+
+export interface BenefitValidationErrors {
+  name?: string;
+  description?: string;
+  cost?: string;
+  endAt?: string;
+  subjectId?: string;
+  icon?: string;
+  category?: string;
+  color?: string;
+  purchaseLimit?: string;
+  purchaseLimitPerStudent?: string;
+}
+
+export interface BenefitPurchaseValidation {
+  canPurchase: boolean;
+  reason?: string;
+}
