+++ src/student/components/studentBenefits/benefitStudentCard/BenefitStudentCard.tsx
@@ -1,12 +1,12 @@
 import { FaShoppingCart, FaCheckCircle, FaHourglassHalf } from "react-icons/fa";
 import { FiXCircle } from "react-icons/fi";
-import type { BenefitStudentResponseInterface } from "../../../../shared/types/Benefits.type";
+import type { BenefitStudentResponseInterface } from "../../../../benefit/types/benefit.types";
 import Button from "../../../../shared/components/Button/ButtonComponent";
 import Card from "../../../../shared/components/Card/CardComponent";
 import Tooltip from "../../../../shared/components/Tooltip/TooltipComponent";
-import BenefitCardContent from "../../../../teacher/components/benefitsView/benefitCardComponent/benefitCardContent/BenefitCardContent";
-import { BENEFIT_STATUS } from "../../../constants/benefitStudent.constants";
-import { validateBenefitPurchase } from "../../../utils/benefitStudent.validation";
+import BenefitCardContent from "../../../../benefit/components/benefitCardContent/BenefitCardContent";
+import { BENEFIT_STATUS } from "../../../../benefit/constants/benefit.constants";
+import { validateBenefitPurchase } from "../../../../benefit/utils/benefit.validation";
 import { useCurrentStudent } from "../../../hooks/useCurrentStudent";
 import styles from "./BenefitStudentCard.module.css";
