+++ src/student/context/benefitStudentContext/BenefitStudentProvider.tsx
@@ -5,7 +5,7 @@ import { BenefitService } from "../../services/benefit/BenefitService";
 import type {
   BenefitStudentResponseInterface,
   BenefitStatsResponse,
-} from "../../../shared/types/Benefits.type";
+} from "../../../benefit/types/benefit.types";
 import type {
   GetPaginated,
   PaginatedData,
