+++ src/teacher/components/benefitsView/benefitCreateView/benefitPreview/BenefitPreview.tsx
@@ -1,7 +1,7 @@
 import { motion } from "framer-motion";
 import Card from "../../../../../shared/components/Card/CardComponent";
-import BenefitCardContent from "../../benefitCardComponent/benefitCardContent/BenefitCardContent";
-import type { CreateBenefitInterface } from "../../../../../shared/types/Benefits.type";
+import BenefitCardContent from "../../../../../benefit/components/benefitCardContent/BenefitCardContent";
+import type { CreateBenefitInterface } from "../../../../../benefit/types/benefit.types";
 import styles from "./BenefitPreview.module.css";
 
 type BenefitPreviewProps = {
