--- src/teacher/constants/benefits.constants.ts
@@ -1,104 +1,0 @@
-import type { IconType } from "react-icons";
-import {
-  FaGraduationCap,
-  FaFileAlt,
-  FaCalendarCheck,
-  FaStar,
-  FaUsers,
-  FaCoins,
-  FaChartLine,
-  FaGift,
-} from "react-icons/fa";
-import type { Category, Color, Icon } from "../../shared/types/Benefits.type";
-
-export const BENEFIT_CATEGORIES: ReadonlyArray<{
-  value: Category;
-  label: string;
-  icon: IconType;
-  color: string;
-  textColor: string;
-}> = [
-  {
-    value: "EVALUACION",
-    label: "Evaluaciones",
-    icon: FaGraduationCap,
-    color: "#007bff",
-    textColor: "#fff",
-  },
-  {
-    value: "TRABAJOS",
-    label: "Trabajos",
-    icon: FaFileAlt,
-    color: "#ff6f3c",
-    textColor: "#fff",
-  },
-  {
-    value: "ASISTENCIA",
-    label: "Asistencia",
-    icon: FaCalendarCheck,
-    color: "#10b981",
-    textColor: "#fff",
-  },
-  {
-    value: "EXTRAS",
-    label: "Extras",
-    icon: FaStar,
-    color: "#8b5cf6",
-    textColor: "#fff",
-  },
-] as const;
-
-export const BENEFIT_ICON_OPTIONS: ReadonlyArray<{
-  value: Icon;
-  label: string;
-  icon: IconType;
-}> = [
-  { value: "EXAM", label: "Examen", icon: FaGraduationCap },
-  { value: "FILE", label: "Archivo", icon: FaFileAlt },
-  { value: "SKIP", label: "Saltar", icon: FaStar },
-  { value: "CALENDAR", label: "Calendario", icon: FaCalendarCheck },
-  { value: "CHAT", label: "Chat", icon: FaUsers },
-  { value: "CLOCK", label: "Reloj", icon: FaCoins },
-  { value: "BOOK", label: "Libro", icon: FaChartLine },
-  { value: "RETRY", label: "Reintentar", icon: FaGift },
-] as const;
-
-export const BENEFIT_COLOR_OPTIONS: ReadonlyArray<{
-  value: Color;
-  color: string;
-}> = [
-  { value: "BLUE", color: "#007bff" },
-  { value: "ORANGE", color: "#ff6f3c" },
-  { value: "LIGHTGREEN", color: "#b9e769" },
-  { value: "EMERALD", color: "#10b981" },
-  { value: "PURPLE", color: "#8b5cf6" },
-  { value: "AMBER", color: "#f59e0b" },
-  { value: "RED", color: "#dc2626" },
-  { value: "GRAY", color: "#6b7280" },
-] as const;
-
-export const BENEFIT_PLACEHOLDERS = {
-  SEARCH_BENEFITS: "Buscar beneficios...",
-  DEFAULT_CATEGORY: "Todas las categorías",
-} as const;
-
-export const BENEFIT_FORM_PLACEHOLDERS = {
-  NAME: "Ingrese el nombre del beneficio",
-  DESCRIPTION:
-    "Describe detalladamente qué obtiene el estudiante con este beneficio",
-  COST: "Ingrese una cantidad de monedas",
-  PURCHASE_LIMIT: "Ingrese un límite de compra",
-  PURCHASE_LIMIT_PER_STUDENT: "Ingrese un límite por estudiante",
-} as const;
-
-export const BENEFIT_VALIDATION = {
-  NAME_MAX_LENGTH: 100,
-  DESCRIPTION_MAX_LENGTH: 1000,
-  MIN_COST: 1,
-} as const;
-
-export const BENEFIT_TOOLTIP = {
-  PURCHASE_LIMIT: "Si no se define, el límite es ilimitado.",
-  PURCHASE_LIMIT_PER_STUDENT:
-    "Si no se define, el límite por estudiante es ilimitado.",
-} as const;
