+++ src/teacher/contexts/benefitsAPIContext/BenefitAPIContext.type.ts
@@ -1,7 +1,7 @@
 import type {
   BenefitResponseInterface,
   CreateBenefitInterface,
-} from "../../../shared/types/Benefits.type";
+} from "../../../benefit/types/benefit.types";
 import type {
   GetPaginated,
   PaginatedData,
